third_party/WebKit/LayoutTests/http/tests/loading/sxg/sxg-usecounter.html

<!DOCTYPE html>
<title>SignedHTTPExchange UseCounter</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  await waitUntilDidFinishLoadForFrame;

  const message = await openSXGInIframeAndWaitForMessage(t, 'resources/sxg-usecounter.sxg');
  assert_equals(message.location, 'https://127.0.0.1:8443/loading/sxg/resources/inner-url.html');
  assert_true(message.is_use_counted);
  assert_false(message.is_fallback);
}, 'SignedHTTPExchange inner response documents are use counted.');

</script>
</body>
